requirements.txt
flask
flask-cors
pymysql
python-dotenv
bcrypt
PyJWT
requests

